+++ NEWS.md
@@ -1,5 +1,10 @@
 # insight 0.18.4
 
+## Changes to functions
+
+* `model_info()` now includes information for `htest` objects from
+  `shapiro.test()` and `bartlett.test()` (will return `$is_variancetest = TRUE`).
+
 ## Bug fixes
 
 * Fixed issue in `get_data()` which did not correctly backtransform to original